% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_folios.R
\encoding{UTF-8}
\name{crear_folios}
\alias{crear_folios}
\title{Crear identificadores unicos para vivienda, hogar y persona}
\usage{
crear_folios(data)
}
\arguments{
\item{data}{Un data.frame o tibble con las variables base (\code{cd_a}, \code{ent}, \code{con}, \code{v_sel}).}
}
\value{
Un data.frame con las columnas \code{folio}, \code{folio2} y opcionalmente \code{folio3}.
}
\description{
Esta funcion genera los identificadores \code{folio}, \code{folio2} y \code{folio3}
basados en las variables clave del diseno de la ENOE.
}
\seealso{
Other procesamiento_enoe:
\code{\link[=.armonizar_sinco_enoe_core]{.armonizar_sinco_enoe_core()}},
\code{\link[=armonizar_carreras_enoe]{armonizar_carreras_enoe()}},
\code{\link[=calcular_desajuste_estadistico]{calcular_desajuste_estadistico()}},
\code{\link[=calcular_desajuste_horizontal]{calcular_desajuste_horizontal()}},
\code{\link[=imputa_ingocup]{imputa_ingocup()}},
\code{\link[=ipc_enoe]{ipc_enoe()}},
\code{\link[=procesar_clases_damian]{procesar_clases_damian()}},
\code{\link[=procesar_contribucion_hogar]{procesar_contribucion_hogar()}},
\code{\link[=procesar_cuidado_extra]{procesar_cuidado_extra()}},
\code{\link[=procesar_estudio_trabajo]{procesar_estudio_trabajo()}},
\code{\link[=procesar_libro1]{procesar_libro1()}},
\code{\link[=procesar_tiempo]{procesar_tiempo()}},
\code{\link[=procesar_variables_enoe]{procesar_variables_enoe()}},
\code{\link[=procesar_vars_hogar]{procesar_vars_hogar()}},
\code{\link[=procesar_vars_laborales]{procesar_vars_laborales()}},
\code{\link[=procesar_vars_sociodemo]{procesar_vars_sociodemo()}},
\code{\link[=reglas_enoe]{reglas_enoe()}},
\code{\link[=sinco2019_to_sinco2011]{sinco2019_to_sinco2011()}}
}
\concept{procesamiento_enoe}
