% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_cuestionarios.R
\encoding{UTF-8}
\name{descargar_cuestionarios}
\alias{descargar_cuestionarios}
\title{Descargar cuestionarios tecnicos de la ENOE}
\usage{
descargar_cuestionarios(
  anio,
  trimestre,
  destino = "cuestionarios",
  sobrescribir = FALSE
)
}
\arguments{
\item{anio}{Ano del trimestre (2005-2024). Debe ser un valor numerico entre 2005 y 2024.}

\item{trimestre}{Numero del trimestre (1-4). Donde 1 = ENE-MAR, 2 = ABR-JUN,
3 = JUL-SEP, 4 = OCT-DIC.}

\item{destino}{Directorio base donde se guardaran los archivos. Por defecto, "cuestionarios".}

\item{sobrescribir}{Logico. ?Deben sobrescribirse los archivos existentes?
(FALSE por defecto).}
}
\value{
Vector invisible con las rutas de los archivos descargados.
}
\description{
Descarga automaticamente los cuestionarios en PDF y los descriptores de archivos (FD) correspondientes
a un trimestre especifico de la ENOE, basandose en la informacion de versiones. Los archivos se almacenan
en una carpeta unica con nombre \verb{docs_\{anio\}_\{trim\}} dentro del directorio especificado.
}
\examples{
\dontrun{
descargar_cuestionarios(2020, 1)
descargar_cuestionarios(2019, 2, sobrescribir = TRUE)
descargar_cuestionarios(2018, 3, destino = "documentacion_enoe")
}
}
\seealso{
Other descarga_documenta_enoe:
\code{\link[=carga_enoe]{carga_enoe()}},
\code{\link[=descarga_enoe]{descarga_enoe()}},
\code{\link[=fusion_enoe]{fusion_enoe()}},
\code{\link[=info_trimestre]{info_trimestre()}}
}
\concept{descarga_documenta_enoe}
