% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.construir_url_enoe}
\alias{.construir_url_enoe}
\title{Funcion interna que determina la estructura de URL apropiada segun el ano y trimestre}
\usage{
.construir_url_enoe(anio, trimestre)
}
\description{
Funcion interna que determina la estructura de URL apropiada segun el ano y trimestre
}
\keyword{internal}
