% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.descargar_zip_enoe}
\alias{.descargar_zip_enoe}
\title{Descarga robusta de archivos ZIP}
\usage{
.descargar_zip_enoe(url, destfile, intentos = 3, timeout = 300)
}
\description{
Descarga robusta de archivos ZIP
}
\keyword{internal}
