% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.estandarizar_ids}
\alias{.estandarizar_ids}
\title{Estandariza variables de identificacion}
\usage{
.estandarizar_ids(df, anio, trimestre)
}
\description{
Estandariza variables de identificacion
}
\keyword{internal}
