% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{.leer_datos_enoe}
\alias{.leer_datos_enoe}
\title{Leer archivos de datos de la ENOE desde directorio descomprimido}
\usage{
.leer_datos_enoe(tabla, unzip_dir, prefijo, anio, trimestre)
}
\description{
Leer archivos de datos de la ENOE desde directorio descomprimido
}
\keyword{internal}
