% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.procesar_etiquetas_enoe}
\alias{.procesar_etiquetas_enoe}
\title{Procesamiento de etiquetas y metadatos}
\usage{
.procesar_etiquetas_enoe(datos, tabla, anio, trimestre, unzip_dir, prefijo)
}
\description{
Procesamiento de etiquetas y metadatos
}
\keyword{internal}
