% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.sustituir_todo_enoe_2022t1}
\alias{.sustituir_todo_enoe_2022t1}
\title{Sustituye los cinco archivos de datos de ENOE 2022T1 con una versión alternativa descargada desde INEGI.}
\usage{
.sustituir_todo_enoe_2022t1(unzip_dir)
}
\description{
Sustituye los cinco archivos de datos de ENOE 2022T1 con una versión alternativa descargada desde INEGI.
}
\keyword{internal}
