% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_contribucion_hogar.R
\name{procesar_contribucion_hogar}
\alias{procesar_contribucion_hogar}
\title{Procesar contribucion economica y de trabajo no remunerado al hogar}
\usage{
procesar_contribucion_hogar(data)
}
\arguments{
\item{data}{Un data frame que contenga, al menos, las variables \code{ingocup_imp},
\code{ipc}, \code{hrsocup}, \code{t_trabajo_hogar_armonizado}, \code{t_total_instrumento}, \code{folio2},
\code{tam_hog} y
\code{fac}.}
}
\value{
Un data frame con variables derivadas sobre contribucion economica y
trabajo no remunerado en el hogar, junto con quintiles ponderados etiquetados.
}
\description{
A partir de variables laborales previamente construidas, genera el ingreso
ocupacional individual deflactado, agregados del hogar, indicadores per capita
y quintiles ponderados de ingreso y trabajo no remunerado.
}
\examples{
\dontrun{
datos <- datos |>
  renoe::procesar_vars_laborales() |>
  renoe::procesar_contribucion_hogar()
}
}
\seealso{
Other procesamiento_enoe:
\code{\link[=.armonizar_sinco_enoe_core]{.armonizar_sinco_enoe_core()}},
\code{\link[=armonizar_carreras_enoe]{armonizar_carreras_enoe()}},
\code{\link[=calcular_desajuste_estadistico]{calcular_desajuste_estadistico()}},
\code{\link[=calcular_desajuste_horizontal]{calcular_desajuste_horizontal()}},
\code{\link[=crear_folios]{crear_folios()}},
\code{\link[=imputa_ingocup]{imputa_ingocup()}},
\code{\link[=ipc_enoe]{ipc_enoe()}},
\code{\link[=procesar_clases_damian]{procesar_clases_damian()}},
\code{\link[=procesar_cuidado_extra]{procesar_cuidado_extra()}},
\code{\link[=procesar_estudio_trabajo]{procesar_estudio_trabajo()}},
\code{\link[=procesar_libro1]{procesar_libro1()}},
\code{\link[=procesar_tiempo]{procesar_tiempo()}},
\code{\link[=procesar_variables_enoe]{procesar_variables_enoe()}},
\code{\link[=procesar_vars_hogar]{procesar_vars_hogar()}},
\code{\link[=procesar_vars_laborales]{procesar_vars_laborales()}},
\code{\link[=procesar_vars_sociodemo]{procesar_vars_sociodemo()}},
\code{\link[=reglas_enoe]{reglas_enoe()}},
\code{\link[=sinco2019_to_sinco2011]{sinco2019_to_sinco2011()}}
}
\concept{procesamiento_enoe}
