% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renoe-package.R
\docType{package}
\name{renoe-package}
\alias{renoe}
\alias{renoe-package}
\title{renoe: Herramientas para trabajar con la ENOE desde 2005}
\description{
Este paquete proporciona funciones para descargar, cargar, fusionar y procesar
datos de la Encuesta Nacional de Ocupación y Empleo (ENOE) de México desde 2005.
}
\details{
Incluye utilidades para imputación de ingresos ocupacionales, procesamiento de
estructura del hogar, uso del tiempo, codificación ocupacional y desajuste
educativo, así como funciones para manejar archivos ZIP, metadatos del INEGI y
formatos como \code{.parquet}, \code{.dta} y \code{.rds}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aniuxa.github.io/renoe}
  \item \url{https://github.com/aniuxa/renoe}
  \item Report bugs at \url{https://github.com/aniuxa/renoe/issues}
}

}
\author{
\strong{Maintainer}: Ana Escoto \email{ana.escoto@politicas.unam.mx} (\href{https://orcid.org/0000-0001-7259-0001}{ORCID})

Authors:
\itemize{
  \item Ana Escoto \email{ana.escoto@politicas.unam.mx} (\href{https://orcid.org/0000-0001-7259-0001}{ORCID})
}

Other contributors:
\itemize{
  \item Gerardo DamiC!n HernC!ndez (\href{https://orcid.org/0009-0002-7604-3886}{ORCID}) [contributor]
  \item Gabriela Cervantes [contributor]
}

}
\keyword{internal}
