% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armoniza_sinco.R
\encoding{UTF-8}
\name{armoniza_sinco}
\alias{armoniza_sinco}
\title{Armonizar códigos SINCO desde CMO o SINCO 2011}
\usage{
armoniza_sinco(data, codigos = NULL)
}
\arguments{
\item{data}{Un data.frame con al menos \code{anio}, \code{trimestre} (como "t1", "t2", etc.) y \code{p3coe}.
También puede contener \code{sinco4d}, \code{pos_ocu}, \code{tue2}.}

\item{codigos}{(Opcional) data.frame de correspondencia CMO–SINCO. Si se omite, se usa la tabla interna del paquete.}
}
\value{
Un data.frame con variables armonizadas: \code{cmo_4d}, \code{sinco4d}, \code{sinco3d}, \code{sinco2d}, \code{sinco1d}.
}
\description{
Esta función construye variables \code{sinco1d}, \code{sinco2d}, \code{sinco3d} y \code{sinco4d} a partir de códigos CMO (anteriores a 2013,
o hasta el segundo trimestre de 2013) o directamente desde SINCO 2011 (desde el tercer trimestre de 2013 en adelante),
utilizando una correspondencia incluida en el paquete y reglas manuales de clasificación cuando no hay equivalencia directa.
}
\examples{
datos <- data.frame(
  anio = c(2011, 2012, 2013),
  trim = c("t1", "t2", "t3"),
  p3coe = c(7121, 4190, 1110),
  pos_ocu = c(1, 2, 1),
  tue2 = c(1, 2, 1)
)
armoniza_sinco(datos)
}
\seealso{
Other procesamiento_enoe: 
\code{\link{crear_folios}()},
\code{\link{imputa_ingocup}()},
\code{\link{ipc_enoe}()},
\code{\link{procesar_tiempo}()},
\code{\link{procesar_variables_enoe}()},
\code{\link{procesar_vars_hogar}()},
\code{\link{procesar_vars_laborales}()},
\code{\link{procesar_vars_sociodemo}()}
}
\concept{procesamiento_enoe}
