% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmo_to_sinco1d.R
\encoding{UTF-8}
\name{cmo_to_sinco1d}
\alias{cmo_to_sinco1d}
\title{Convertir c칩digos CMO a clasificaci칩n SINCO 1d (nivel agregado)}
\usage{
cmo_to_sinco1d(data)
}
\arguments{
\item{data}{Un data frame con variables \code{p3_coe}, \code{anio}, \code{trim} y preferentemente \code{pos_ocu}, \code{tue2}.}
}
\value{
El data frame original con la variable \code{sinco1d} agregada.
}
\description{
Aplica mapeo manual para a침os anteriores a 2013 o hasta el segundo trimestre de 2013,
con base en \code{p3_coe}, usando reglas por \code{cmo2d}, \code{cmo3d} y c칩digos completos.
}
