% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_folios.R
\encoding{UTF-8}
\name{crear_folios}
\alias{crear_folios}
\title{Crear identificadores únicos para vivienda, hogar y persona}
\usage{
crear_folios(data)
}
\arguments{
\item{data}{Un data.frame o tibble con las variables base (\code{cd_a}, \code{ent}, \code{con}, \code{v_sel}).}
}
\value{
Un data.frame con las columnas \code{folio}, \code{folio2} y opcionalmente \code{folio3}.
}
\description{
Esta función genera los identificadores \code{folio}, \code{folio2} y \code{folio3}
basados en las variables clave del diseño de la ENOE.
}
