% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.construir_url_enoe}
\alias{.construir_url_enoe}
\title{Función interna que determina la estructura de URL apropiada según el año y trimestre}
\usage{
.construir_url_enoe(anio, trimestre)
}
\description{
Función interna que determina la estructura de URL apropiada según el año y trimestre
}
\keyword{internal}
