% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_trimestre.R
\encoding{UTF-8}
\name{info_trimestre}
\alias{info_trimestre}
\title{Obtener metadatos de versiones de cuestionarios ENOE por trimestre}
\usage{
info_trimestre(anio, trimestre)
}
\arguments{
\item{anio}{Año del trimestre (2005-2024). Debe ser un valor numérico entre 2005 y 2025.}

\item{trimestre}{Núemero del trimestre (1-4). Donde 1 = ENE-MAR, 2 = ABR-JUN,
3 = JUL-SEP, 4 = OCT-DIC.}
}
\value{
Un data.frame con 6 columnas:
\describe{
\item{trimestre}{Código del trimestre (ej. "t105")}
\item{coe_tipo}{Tipo de cuestionario COE ("ampliado" o "basico")}
\item{coe_v}{Versión del cuestionario COE (ej. "v1", "v2")}
\item{sdem_v}{Versión del cuestionario SDEM}
\item{fd}{Versión del file descriptor}
\item{encoding}{Encoding recomendado para los archivos}
}
Retorna NULL si no se encuentra información para el trimestre especificado.
}
\description{
Consulta la informacipib sobre qué versiones de cuestionarios (COE, SDEM, FD)
corresponden a un trimestre específico de la ENOE.
}
\examples{
# Consultar información para el primer trimestre de 2020
info_trimestre(2020, 1)

# Consultar información para el tercer trimestre de 2015
info_trimestre(2015, 3)
}
