% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc_enoe.R
\name{ipc_enoe}
\alias{ipc_enoe}
\title{Añadir IPC al conjunto fusionado de ENOE}
\usage{
ipc_enoe(datos_fusionados, anio, trimestre)
}
\arguments{
\item{datos_fusionados}{Un data.frame ya fusionado con \code{fusion_enoe()}.}

\item{anio}{Año del trimestre (numérico).}

\item{trimestre}{Trimestre (1 a 4).}
}
\value{
El mismo \code{data.frame} con una nueva columna \code{ipc}.
}
\description{
Esta función agrega una columna llamada \code{ipc} al objeto fusionado de la ENOE,
correspondiente al promedio trimestral del Índice de Precios al Consumidor (IPC).
}
\details{
El archivo \code{ipc.rds} debe estar ubicado en \verb{inst/extdata/} y contener
las columnas \code{anio}, \code{trim} e \code{ipc}.
}
\examples{
\dontrun{
datos <- fusion_enoe(2023, 1)
datos <- ipc_enoe(datos, 2023, 1)
}
}
\seealso{
Other procesamiento_enoe: 
\code{\link{armoniza_sinco}()},
\code{\link{crear_folios}()},
\code{\link{imputa_ingocup}()},
\code{\link{procesar_tiempo}()},
\code{\link{procesar_variables_enoe}()},
\code{\link{procesar_vars_hogar}()},
\code{\link{procesar_vars_laborales}()},
\code{\link{procesar_vars_sociodemo}()}
}
\concept{procesamiento_enoe}
