% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_contribucion_hogar.R
\name{procesar_contribucion_hogar}
\alias{procesar_contribucion_hogar}
\title{Procesar contribución económica y de trabajo no remunerado al hogar}
\usage{
procesar_contribucion_hogar(data)
}
\arguments{
\item{data}{Un data frame que contenga, al menos, las variables \code{ingocup_imp},
\code{ipc}, \code{t_total_hrs0}, \code{t_total_hrs}, \code{folio2}, \code{tam_hog} y \code{fac}.}
}
\value{
Un data frame con variables derivadas sobre contribución económica y
trabajo no remunerado en el hogar, junto con quintiles ponderados etiquetados.
}
\description{
A partir de variables laborales previamente construidas, genera el ingreso
ocupacional individual deflactado, agregados del hogar, indicadores per cápita
y quintiles ponderados de ingreso y trabajo no remunerado.
}
\examples{
\dontrun{
datos <- datos |>
  renoe::procesar_vars_laborales() |>
  renoe::procesar_contribucion_hogar()
}
}
\seealso{
Other procesamiento_enoe: 
\code{\link{armoniza_sinco}()},
\code{\link{crear_folios}()},
\code{\link{imputa_ingocup}()},
\code{\link{ipc_enoe}()},
\code{\link{procesar_tiempo}()},
\code{\link{procesar_variables_enoe}()},
\code{\link{procesar_vars_hogar}()},
\code{\link{procesar_vars_laborales}()},
\code{\link{procesar_vars_sociodemo}()}
}
\concept{procesamiento_enoe}
