% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_variables_enoe.R
\encoding{UTF-8}
\name{procesar_variables_enoe}
\alias{procesar_variables_enoe}
\title{Procesar variables sociodemográficas y de estructura del hogar}
\usage{
procesar_variables_enoe(data)
}
\arguments{
\item{data}{Un data frame con las tablas fusionadas de la ENOE (por ejemplo, salida de \code{fusion_enoe()}).}
}
\value{
Un data frame con variables sociodemográficas generales y de estructura del hogar agregadas.
}
\description{
Esta función aplica en cadena \code{procesar_vars_sociodemo()} y \code{procesar_vars_hogar()} para
crear variables derivadas clave en el análisis de datos de la ENOE. Incluye transformación
de edad, identificación de parentescos, clasificación de hogares y cálculo de tasas de dependencia.
}
\examples{
\dontrun{
datos <- fusion_enoe(2022, 1)
datos_proc <- procesar_variables_enoe(datos)

dplyr::glimpse(datos_proc)
table(datos_proc$tipo_hog_lab, useNA = "always")
}
}
\seealso{
\code{\link[=procesar_vars_sociodemo]{procesar_vars_sociodemo()}}, \code{\link[=procesar_vars_hogar]{procesar_vars_hogar()}}
}
