% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_vars_hogar.R
\encoding{UTF-8}
\name{procesar_vars_hogar}
\alias{procesar_vars_hogar}
\title{Procesar variables de estructura del hogar en la ENOE}
\usage{
procesar_vars_hogar(data)
}
\arguments{
\item{data}{Un data frame con variables como \code{par_c}, \code{edad}, \code{sexo}, \code{folio2}, previamente procesadas por \code{procesar_vars_sociodemo()}.}
}
\value{
Un data frame con variables nuevas:
\describe{
\item{relative}{Clasificaci칩n del parentesco con respecto al jefe}
\item{family, familyt, tipo_hog, tipo_hog_lab}{Clasificaciones del tipo de hogar}
\item{tam_hog, t_dep1, t_dep2, t_dep3}{Tama침o del hogar y tasas de dependencia}
}
}
\description{
Calcula el tipo de parentesco, clasificaciones de hogares (familiares, extensos, compuestos, etc.)
y tasas de dependencia. Esta funci칩n requiere que las variables sociodemogr치ficas ya hayan sido creadas.
}
\examples{
\dontrun{
datos <- fusion_enoe(2021, 1)
datos <- procesar_vars_sociodemo(datos)
datos <- procesar_vars_hogar(datos)
table(datos$tipo_hog_lab, useNA = "always")
}
}
