% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_vars_laborales.R
\name{procesar_vars_laborales}
\alias{procesar_vars_laborales}
\title{Procesar variables de análisis laboral y desajuste educativo}
\usage{
procesar_vars_laborales(data)
}
\arguments{
\item{data}{Un data.frame con variables como:
\itemize{
\item \code{anio}, \code{trimestre}: año y trimestre de la entrevista
\item \code{coe_tipo}: tipo de cuestionario (\code{"ampliado"} o \code{"basico"})
\item \code{p3coe}: código ocupacional
\item \code{cs_p13_1}, \code{anios_es}: nivel educativo alcanzado en categorías o años
\item \code{clase2}: clase de actividad económica
\item \code{pos_ocu}, \code{tue2}: posición en la ocupación y tipo de unidad económica
\item \code{p2_4}: experiencia laboral previa
\item \code{p3i}, \code{p3j}, \code{p3j1}, \code{p3k1}: variables sobre tipo de contrato
}}
}
\value{
Un data.frame con las variables originales y nuevas columnas:
\itemize{
\item \code{sinco1d}, \code{sinco2d}, \code{sinco3d}, \code{sinco4d}
\item \code{skill_level}, \code{skill_actual}
\item \code{mismatch}, \code{mismatch2}
\item \code{nunca_trabajo}, \code{status_seq}
\item \code{contrato0}, \code{contrato1}, \code{temporal}, \code{temporal_seq}
}
}
\description{
Esta función genera variables clasificatorias relacionadas con la ocupación, nivel educativo
y desajuste entre ambos, a partir de los códigos de ocupación (\code{p3coe}) y nivel educativo alcanzado (\code{cs_p13_1}).
Internamente armoniza los códigos SINCO (1, 2, 3 y 4 dígitos) usando correspondencias con códigos CMO
y reglas auxiliares. También clasifica el nivel de habilidad de la ocupación, el nivel educativo alcanzado
y construye indicadores de sobrecalificación, ajuste o subcalificación laboral.
}
\details{
Además, genera variables relacionadas con la experiencia previa (\code{nunca_trabajo}), estatus laboral combinado (\code{status_seq})
y características contractuales (\code{contrato0}, \code{contrato1}, \code{temporal}, \code{temporal_seq}) según el tipo de cuestionario.
}
\seealso{
Other procesamiento_enoe: 
\code{\link{armoniza_sinco}()},
\code{\link{crear_folios}()},
\code{\link{imputa_ingocup}()},
\code{\link{ipc_enoe}()},
\code{\link{procesar_tiempo}()},
\code{\link{procesar_variables_enoe}()},
\code{\link{procesar_vars_hogar}()},
\code{\link{procesar_vars_sociodemo}()}
}
\concept{procesamiento_enoe}
